[Infra][Optimize] create or reuse venv if no deps installed in current python env
The jni generation process depends on the `PyYAML` module, the script will try to install the module when it is not found in the current env, but it leads some issues:

1. pollute the global python env
2. may face failure due to PEP-668

This patch fixes this by providing a new script `vpython` to create a venv if it is not present in the current python env, and will check & reuse the previous venv. By addtion, the script checks & installs the `requests` module needed later.