[Optimize] Add a switch for controlling the behavior of platform layer's LayoutNodeManager.
In Lynx 3.2, we added thread control to the platform layer entry of LayoutNodeManager to ensure that all layout-related methods without return values are only called on the Layout thread. However, businesses that rely on unsafe thread-safe behaviors may produce breaking changes, so this commit introduces a switch to roll back the behavior.

issue: m-6271346517