[Optimize][Text] Set enableNewClipMode as true by default.
The enableNewClipMode is used to control the text clip mode, align the double-ended clip mode when turned on, and support word-break:break-all on Android. This switch was added in version 2.10 and defaults to false. It defaults to true when targetSdkVersion is greater than or equal to 2.10.

Set enableNewClipMode as true by default. We will delete this switch in future versions.

issue:m-6546469094