[BugFix][List] Fix list error info with diff result.
If we use std::ostringstream and invoke the constructor with parameter std::string, we should invoke seekp() function to set the next char position in output sequence, otherwise the char sequence will be replaced by the next invoking operator<<(). So in this case, we invoke default constructor directly.

issue: m-6083156228