[BugFix] Align mouse enter/leave event with W3C spec
According to the W3C specification, the `mouseenter` and `mouseleave` events do not bubble. Therefore, bubbling needs to be disabled.
The mouse event also should not affected by the `z-index` attribute, so API `GetAncestorElements` is added to get the original list of elements that is not affected by z-index.

issue:m-6577356292