[Optimize][Event] Add `consumed` state to `DispatchEventResult`
Sometimes we want to check whether the event is really consumed. For example, long press and tap events may be generated at the same time, in which case, we should trigger the tap events only if the long press is not consumed on the front-end.

To achieve this, we convert `DispatchEventResult` from an enum to a struct and add a `consumed` field.

issue: m-6577319379

AutoLand: release/3.2