[Optimize] Rename CSSPatching to StyleResolver and remove unnecessary member variables
According to previous TOOD, the CSSPatching class should be renamed to StyleResolver. In this commit, we also make StyleResolver a class without any member variables so that we do not need to synchronize any flag or pointer variables across multiple objects. The sizeof(FiberElement) is reduced by 24 bytes.

issue:m-6383995029