[Optimize][Element] Track Fiber Element Allocation size
In the current implementation, the virtual machine does not track the amount of memory allocated for fiber elements, leading to an increase in out-of-memory (OOM) crashes under tracing garbage collection (GC) mode.

This merge request addresses the issue by reporting the approximate size of a new FiberElement to the virtual machine at the end of pixel pipeline to reduce GC impact while running pixel pipeline, enabling the garbage collection to be triggered with an appropriate strategy.

issue:m-0

AutoSubmit: true