[Optimize][Harmony] Fix crash caused by asynchronous release of PerformanceControllerHarmonyJSWrapper.
This MR is to fix crash caused by asynchronous release of PerformanceControllerHarmonyJSWrapper.

After the PerformanceControllerHarmonyJSWrapper is released on the Report thread, an exception may occur when accessed from the UI thread.
The solution is to create, access, and release the PerformanceControllerHarmonyJSWrapper on the UI thread.

issue: m-6741122532
AutoSubmit: True
AutoLand:release/3.4