[BugFix][Android][iOS] Fix the problem of reported event loss.
This MR is to fix the problem of reported event loss.

Since LynxService may be registered late, it cannot be obtained in the constructor of LynxEventReporter, resulting in a failure in reporting the event. The solution is to delay obtaining LynxService until method `LynxEventReporter.handleEvent`.

issue: m-6741201001
AutoSubmit:true
AutoLand:release/3.4